// interfaces/posProfile/iPOSProfile.ts
import { iPOSPaymentMethod } from "./iPOSPaymentMethod";
import { iPOSProfileUser } from "./iPOSProfileUser";

export interface iPOSProfile {
    account_for_change_amount: string | null;
    allow_discount_change: number;
    allow_rate_change: number;
    applicable_for_users: iPOSProfileUser[];
    apply_discount_on: string;
    auto_add_item_to_cart: number;
    campaign: string | null;
    company: string;
    company_address: string | null;
    cost_center: string | null;
    country: string;
    creation: string;
    currency: string;
    customer: string | null;
    customer_groups: string[];
    disable_rounded_total: number;
    disabled: number;
    docstatus: number;
    doctype: string;
    expense_account: string | null;
    hide_images: number;
    hide_unavailable_items: number;
    idx: number;
    ignore_pricing_rule: number;
    income_account: string | null;
    item_groups: string[];
    letter_head: string | null;
    modified: string;
    modified_by: string;
    name: string;
    owner: string;
    payments: iPOSPaymentMethod[];
    print_format: string | null;
    select_print_heading: string | null;
    selling_price_list: string;
    tax_category: string | null;
    taxes_and_charges: string;
    tc_name: string | null;
    update_stock: number;
    validate_stock_on_save: number;
    warehouse: string;
    write_off_account: string;
    write_off_cost_center: string;
    write_off_limit: number;
  }